Fix update-deps.sh
I misunderstood how `godep update` works, and I still think the
workflow of the tool is weird, but this at least does what it says on
the label, now.  The `go get -u` is required to pull down the updated
dependencies; `godep update` just vendors the dependency and updates
Godeps.json.  `godep save` is still required to pull in *new*
dependencies.