Add support for Consul TTLs and multiple checks.
This may potentially close #57, depending on how the original author wishes to
solve the problem. On each refresh, Registrator will write the current timestamp
into the check's notes.

Further, this change also introduces the ability to have multiple health checks
types. For example, a TTL may be used in conjunction with a script. This allows
for a more reliable service discovery mechanism, reducing the risk of services
reclaiming old ports. These changes also open the door for a simple, native HTTP
check integration with Consul. Because Registrator doesn't currently have an API
to store the upstream state of the service (i.e. remote check IDs), the Consul
adapter filters the defined checks into an ordered list, which it then
correlates to the check ID in Consul based on index. Consul's API implementation
suggests this is safe, however, this behaviour is not documented.

Sadly, this breaks compatibility with Consul 0.4.x, which only supported one
service check per service.