Add service cleanup to consul registry
When stopping registrator uncleanly (either by SIGKILL or when consul is not
available) it has no chance to cleanup its registered services with consul.
This patch adds a Cleanup method to the Registry interface and implements it
for consul to find its service again (ID prefix with "registrator-") and remove
those which are not valid anymore.