Add '-deregister-on-stop' option
Currently registrator deregisters the service once the container dies,
but in some cases that might be too late. The -deregister-on-stop option
tells registrator to deregister the service once docker receives either
the stop or kill commands, but before the container actually dies.

This gives the service in the container a chance to clean up after
itself after receiving a SIGTERM, without the chance that it will
receive any new requests before it actually dies.