Service aliases
This commit makes it possible for a given container to advertise
multiple services using aliases.

For example, the following config:

```
SERVICE_NAME=a
SERVICE_ALIASES=b,c
```

will result in 3 services being registered: a, b, and c. All three
pointing to the same container.

Those changes should be compatible with ports, and tags.