Add go.{mod/sum} files
As part of upgrading the codebase to golang v1.16, add the go.mod
and go.sum files in order to leverage the latest go modules features.

This will also signal that the repository is go module given the
presence of the go.mod on the root of the project.

The files was added using the `go mod init` command.