Rebuild service state during sync
tenant-maxscale often fails its Consul TCP health check despite being
able to telnet to its ports, which in turn fails builds when bringing up
the cleanroom. When in this state, Consul displays its IP as 172.18.0.1,
which is incorrect. It is believed this issue is not unique to
tenant-maxscale.

When debugging Registrator, Bridge.service's tenant-maxscale Service.IP
was observed to be "" (the same as Consul) while Docker had a value for
the IP. It is currently unknown how Registrator got into this state.
Nevertheless, because the service information is cached, the service is
never re-registered with Consul using the correct IP.

- Bumped VERSION.

- Added debugging run configuration.

- Updated Bridge.Sync to rebuild service state.

- Always log when syncing.

Testing
=======
Although this issue is not consistently reproduceable, performing a
handful of builds in a row typically reproduces it.  This change was
tested by executing `mvn clean install` on the cleanroom module of the
cloud-work-service repo 10x in a row, and ensuring Consul and
Registrator were not restarted at any time.

Building
========
While investigating this issue, building is being performed namually
via:
```
tag="registry.imanage.com/unverified/davew/registrator:$(cat VERSION)" && docker build --tag ${tag} . && docker push ${tag}
```

Running
=======
Recommended options are `-cleanup -internal -resync 60`.
- `-cleanup` is only valid with `-resync`.
- With this change, `-resync` not only lists all containers, it inspects
  each one. Because of this, it is thought that less frequent resyncs
may be better.

To Do
=====
- [ ] Upgrade Docker build
- [ ] Upgrade Alpine
- [ ] Upgrade Go
- [ ] Migrate from [dep](https://golang.github.io/dep/docs/introduction.html) to Go modules
- [ ] Upgrade dependencies (see Gopkg.toml)